First implementation of AssociateAction (successor of MetaMaskImport).
