+++ src/pages/TokenDetails.vue
@@ -52,17 +52,16 @@
         </div>
 
         <div v-if="ethereumAddress" id="evmAddress"
-             class="headline-grid is-align-items-baseline h-is-property-text mt-2" style="word-break: keep-all">
-          <div class="has-text-weight-light">EVM Address:</div>
-          <div class="is-flex is-align-items-baseline">
-            <EVMAddress class="mr-3" :show-id="false" :has-custom-font="true" :address="ethereumAddress"/>
-            <MetaMaskImport v-if="connectedToMetamask && isSmallScreen" :analyzer="tokenAnalyzer"/>
+               class="headline-grid is-align-items-baseline h-is-property-text mt-2" style="word-break: keep-all">
+              <div class="has-text-weight-light">EVM Address:</div>
+              <div class="is-flex is-align-items-baseline">
+                  <EVMAddress class="mr-3" :show-id="false" :has-custom-font="true" :address="ethereumAddress"/>
+              </div>
           </div>
-        </div>
-        <div v-if="ethereumAddress && connectedToMetamask && !isSmallScreen" class="mt-2 h-is-property-text">
-          <MetaMaskImport :analyzer="tokenAnalyzer"/>
-        </div>
+      </template>
 
+      <template v-slot:control>
+        <AssociateAction v-if="walletConnected" :analyzer="tokenAnalyzer"/>
       </template>
 
       <template v-slot:content>
@@ -310,7 +309,7 @@ import DashboardCard from "@/components/DashboardCard.vue";
 import BlobValue from "@/components/values/BlobValue.vue";
 import TokenAmount from "@/components/values/TokenAmount.vue";
 import Footer from "@/components/Footer.vue";
-import MetaMaskImport from "@/components/token/MetaMaskImport.vue";
+import AssociateAction from "@/components/token/AssociateAction.vue";
 import {EntityID} from "@/utils/EntityID";
 import Property from "@/components/Property.vue";
 import NotificationBanner from "@/components/NotificationBanner.vue";
@@ -345,7 +344,7 @@ export default defineComponent({
     AccountLink,
     NotificationBanner,
     Property,
-    MetaMaskImport,
+    AssociateAction,
     Footer,
     BlobValue,
     DashboardCard,
@@ -425,8 +424,8 @@ export default defineComponent({
     onMounted(() => nftHolderTableController.mount())
     onBeforeUnmount(() => nftHolderTableController.unmount())
 
-    const connectedToMetamask = computed(
-        () => walletManager.isMetamaskWallet.value && walletManager.connected.value)
+    const walletConnected = computed(
+        () => walletManager.connected.value)
 
     return {
       isSmallScreen,
@@ -446,7 +445,7 @@ export default defineComponent({
       parseBigIntString,
       tokenAnalyzer,
       ethereumAddress: tokenAnalyzer.ethereumAddress,
-      connectedToMetamask,
+      walletConnected,
       tokenBalanceTableController,
       nftHolderTableController,
     }
